refactor: moves home layout styles to own file
