+++ components/ContentBlock/index.tsx
@@ -35,7 +35,7 @@ export default function ContentBlock({
         <h3 className={cx(heveticaNueuCondensedBold.className)}>
           {sub_heading}
         </h3>
-        <TextBlocksRenderer content={content} />
+        <TextBlocksRenderer text_content={content} />
 
         {button && (
           <LinkButton
